Updated comparison function
I updated the function of comparison: before, when the agent did not have an utterance for the topic, it returned NIL, which was treated as an utterance, resulting in a high level of global coherence at the very beginning of the experiment, as most agents just uttered NIL. Now NIL is discarded. We look for the most common non-nil utterance and then calculate the ratio of its frequency over the population.