run-batch now creates experiment on copy of config
In run-batch a series of experiments is instantiated. Each instance is now created on a copy of the experiment-parameters (i.e. configuration). This was introduced for experiments in which the configuration can change over time. Making a copy each time ensures that every experiment in the series starts with the intended initial configuration. Previously, changes made to the configuration of experiment 1 would persist when experiment 2 was started. Experiment 2 would therefore start with the configuration that experiment 1 finished with (as opposed to the configuration that experiment 1 started with).